Moved user-settings
